Fix/crud ordenes de compra

+++ pages/api/articulos/delete.ts
@@ -3,6 +3,8 @@ import { Database, open } from "sqlite";
 import { NextApiRequest, NextApiResponse } from "next";
 
 let db: Database<sqlite3.Database, sqlite3.Statement> | null = null;
+const PENDIENTE = 'Pendiente'
+const ENVIADAS = 'Enviada'
 
 export default async function handler(
     req: NextApiRequest,
@@ -20,29 +22,45 @@ export default async function handler(
         }
 
         const {id}:{id:Number} = req.body;
+
         !id && res.status(400).json({ message: "id is required" });
 
-        //borra tambien todos los precios de venta asociados al articulo
         const articulo = await db.get(
             `SELECT * FROM Articulo 
             join Articulo_Precio_Venta on Articulo.id = Articulo_Precio_Venta.articulo_id
             where Articulo.id = ?`,
             [id]
             );
+
         if (!articulo) {
-            return res.status(404).json({ message: `Articulo ${id} not found` });
+            return res.status(404).json({ message: `Articulo ${id} no encontrado` });
+        }
+
+        //verific aque no existan ordenens pendientes ni enviadas, tiene que estar canceladas o recibidas, las ordenes de compran están relacionadas con articulo_proveedor, que a su vez está relacionado con articulo
+        const orders = await db.all(
+            `SELECT * FROM Articulo 
+            join Articulo_Proveedor on Articulo.id = Articulo_Proveedor.articulo_id
+            join Orden_Compra on Articulo_Proveedor.id = Orden_Compra.articulo_proveedor_id
+            join Orden_Compra_Estado on Orden_Compra.id = Orden_Compra_Estado.orden_compra_id
+            where Articulo.id = ? and Orden_Compra_Estado.estado in (?,?)`,
+            [id, PENDIENTE, ENVIADAS]
+            );
+
+        if (orders && orders.length > 0) {
+            return res.status(400).json({ message: `El articulo tiene ordenes en curso` });
         }
 
         const result = await db.run(
             "DELETE FROM Articulo WHERE id = ?",
             [id]
         );
+
         const result2 = await db.run(
             "DELETE FROM Articulo_Precio_Venta WHERE articulo_id = ?",
             [id]
         );
-        
-        return res.status(200).json({ message: `Articulo ${id} deleted` });
+
+        return res.status(200).json({ message: `El articulo ha sido eliminado` });
     }
     catch (error: any) {
         return res.status(500).json({ message: error.message });
